% Generated by roxygen2 (4.0.2): do not edit by hand
\name{inv.vec}
\alias{inv.vec}
\title{Generate a permutation of a given length having a given number of inversions}
\usage{
inv.vec(n, k, bw.ok = TRUE)
}
\arguments{
\item{n}{A positive integer.}

\item{k}{A nonnegative integer at most \code{n}.}

\item{bw.ok}{Logical; whether to allow exchanging large values of \code{k} for their duals to improve efficiency.}
}
\value{
A permutation of \code{n} having \code{k} (non-arbitrary) inversions.
}
\description{
Generate a permutation of a given length having a given number of inversions
}
\examples{
inv.vec(n = 5, k = choose(5, 2))
vec.inv(n = 5, k = 7)
vec.inv(n = 5, k = 7, bw.ok = FALSE)
}

