% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inv-vec.R
\name{inv_vec}
\alias{inv_vec}
\title{Generate a permutation of a given length having a given number of inversions}
\usage{
inv_vec(n, k, allow.dual = TRUE)
}
\arguments{
\item{n}{A positive integer.}

\item{k}{A nonnegative integer at most \code{n}.}

\item{allow.dual}{Logical; whether to allow exchanging large values of \code{k}
for their duals to improve efficiency.}
}
\value{
A permutation of \code{n} having \code{k} (non-arbitrary) inversions.
}
\description{
Generate a permutation of a given length having a given number of inversions
}
\examples{
inv_vec(n = 5, k = choose(5, 2))
inv_vec(n = 5, k = 7)
inv_vec(n = 5, k = 7, allow.dual = FALSE)
}
