% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mahonians}
\alias{mahonians}
\title{Compute the Mahonian numbers.}
\usage{
mahonians(n)
}
\arguments{
\item{n}{A positive integer.}
}
\value{
The number of permutations of \code{n} having \code{k} inversions, as \code{k} ranges from 0 to \code{n}.
}
\description{
Compute the Mahonian numbers.
}
\examples{
mahonians(1)
mahonians(5)
}

