% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mahonians-recursive.R
\name{mahonians_recursive}
\alias{mahonians_recursive}
\title{Compute the Mahonian numbers recursively (very slow for n > 8).}
\usage{
mahonians_recursive(n)
}
\arguments{
\item{n}{A positive integer.}
}
\value{
The number of permutations of \code{n} having \code{k} inversions, as \code{k}
ranges from 0 to \code{n}.
}
\description{
Compute the Mahonian numbers recursively (very slow for n > 8).
}
\examples{
mahonians_recursive(1)
mahonians_recursive(5)
}
