% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tau.a.crit}
\alias{tau.a.crit}
\title{Find the critical test statistic for a given number of objects and type-I error.}
\usage{
tau.a.crit(n, alpha, stat = "tau", ...)
}
\arguments{
\item{n}{A positive integer.}

\item{alpha}{A vector of probabilities (preferably below .5).}

\item{stat}{A character string; the name of the desired element returned by tau.a.p.}
}
\value{
For each entry of alpha, the least extreme test statistic whose p-value is smaller.
}
\description{
Find the critical test statistic for a given number of objects and type-I error.
}
\examples{
tau.a.crit(5, 10 ^ (-1:-3), stat = 'k')
}

