% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tau.crit.table}
\alias{tau.crit.table}
\title{Rapidly generate a table of critical values of tau or an equivalent test statistic.}
\usage{
tau.crit.table(n, alpha, incl.len = TRUE, stat = "tau")
}
\arguments{
\item{n}{A vector of positive integers.}

\item{alpha}{A vector of probabilities (preferably below .5).}

\item{incl.len}{Logical; whether to include a column of values of \code{n} (default to \code{FALSE}; the rows are named by the values of \code{n} in any case).}

\item{stat}{A character string; any of several recognized test statistics (\code{'tau'}, \code{'concord*'}, , \code{'discord*'}, \code{'P'}, \code{'Q'}, \code{'K'}, \code{'k'}, \code{'inv'}, \code{'S'}).}
}
\value{
The number of permutations of \code{n} having \code{k} inversions, as \code{k} ranges from 0 to \code{n}.
}
\description{
Rapidly generate a table of critical values of tau or an equivalent test statistic.
}
\examples{
tau.crit.table(n = 1:5, alpha = 10 ^ (-1:-3), incl.len = TRUE, stat = 'inv')
tau.crit.table(n = 1:5, alpha = 10 ^ (-1:-3), incl.len = TRUE, stat = 'tau')
}

