% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tau-a-crit.R
\name{tau_a_crit}
\alias{tau_a_crit}
\title{Find the critical test statistic for a given number of objects and type-I
error.}
\usage{
tau_a_crit(n, alpha, stat = "tau", ...)
}
\arguments{
\item{n}{A positive integer.}

\item{alpha}{A vector of probabilities (preferably below .5).}

\item{stat}{A character string; the name of the desired element returned by
\code{\link[=tau_a_p]{tau_a_p()}}.}

\item{...}{Additional parameters passed to \code{\link[=tau_a_p]{tau_a_p()}}.}
}
\value{
For each entry of alpha, the least extreme test statistic whose
p-value is smaller.
}
\description{
Find the critical test statistic for a given number of objects and type-I
error.
}
\examples{
tau_a_crit(5, 10 ^ (-1:-3), stat = "k")
}
