% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tau-a-p.R
\name{tau_a_p}
\alias{tau_a_p}
\title{Tabulate tau and p-values for a given length and number of inversions.}
\usage{
tau_a_p(
  n,
  k = floor((choose(n, 2)/2)):0,
  sided = 1,
  tau.fn = stats::cor.test,
  tau.part = "estimate",
  p.part = "p.value",
  ...
)
}
\arguments{
\item{n}{A positive integer.}

\item{k}{A vector of nonnegative integers at most \code{n}.}

\item{sided}{Number of sides to a test; either 1 or 0.}

\item{tau.fn}{A function that computes elements tau and p-value from a pair of vectors.}

\item{tau.part}{The name of the tau element of the value of tau.fn.}

\item{p.part}{The name of the p-value element of the value of tau.fn.}

\item{...}{Additional parameters passed to \code{tau.fn}.}
}
\value{
For each entry of k, the corresponding value of tau and associated p-value.
}
\description{
Tabulate tau and p-values for a given length and number of inversions.
}
\examples{
tau_a_p(5, 0:5, sided = 1)
tau_a_p(5, 5:10, sided = 1)
tau_a_p(5, 0:5, tau.fn = Kendall::Kendall, tau.part = "tau", p.part = "sl")
}
