% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tau.crit.tab.R
\name{tau.crit.tab}
\alias{tau.crit.tab}
\title{Tabulate critical values for tau and other test statistics via backward use of correlation test functions.}
\usage{
tau.crit.tab(n, alpha, ...)
}
\arguments{
\item{n}{A vector of positive integers.}

\item{alpha}{A vector of probabilities (preferably below .5).}
}
\value{
For each pair of values from \code{n} and \code{alpha}, the critical value of the desired test statistic for rankings of \code{1:n}.
}
\description{
Tabulate critical values for tau and other test statistics via backward use of correlation test functions.
}
\examples{
tau.crit.tab(n = seq(10, 30, 10), alpha = 10 ^ (-1:-2))
}

